--- assets/js/gameController.js
@@ -1,20 +1,0 @@
-(()=>{
-
-    let currentPlayer;
-    let winner = false;
-
-    // game setup
-    // gameloop
-
-
-})();
-
-function gameLoop(){
-    while(!winner){
-        // render board
-        // decide turns
-        // player play
-        // board evalution
-        
-    }
-}
