+++ assets/js/lib/gameController.js
@@ -1,0 +1,18 @@
+// Handles Game Logic & Current Player Turn
+
+(() => {
+  let currentPlayer;
+  let winner = false;
+
+  // game setup
+  // gameloop
+})();
+
+function gameLoop () {
+  while (!winner) {
+    // render board
+    // decide turns
+    // player play
+    // board evalution
+  }
+}
