+++ assets/js/lib/players.js
@@ -1,0 +1,10 @@
+// eslint-disable-next-line no-undef
+class Player {
+  create = (name, token) => {
+    const getName = () => name;
+    const getToken = () => token;
+    return { getName, getToken };
+  }
+}
+
+module.exports = Player;
