+++ index.html
@@ -1,19 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="assets/css/style.css">
-    <title>TIC TAC TOE</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="assets/css/style.css">
+  <title>TIC TAC TOE</title>
 </head>
 <body>
-    <div class="container">
-        <h1>TIC TAC TOE</h1>
-        <div id="board" class="boardgame"></div>
-    </div>
-    <script  src="assets/js/ui.js"></script>
-    <script src="assets/js/main.js"></script>
-    
+  <div class="container">
+    <h1>TIC TAC TOE</h1>
+    <div id="board" class="boardgame"></div>
+  </div>
+  <script src="assets/js/main.js"></script>
 </body>
 </html>
